% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/denseSpaceVisualization.R
\name{denseSpaceVisualization}
\alias{denseSpaceVisualization}
\title{Dense Space visualization}
\usage{
denseSpaceVisualization(
  group = c("sudo", "docker"),
  scratch.folder,
  file,
  separator,
  cloutput,
  perplexity = 30,
  n_iter = 250,
  n_neighbors = 15,
  min_dist = 0.1,
  tsneplot = c(0, 1),
  umapplot = c(0, 1)
)
}
\arguments{
\item{group, }{a character string. Two options: sudo or docker, depending to which group the user belongs}

\item{scratch.folder, }{a character string indicating the path of the scratch folder}

\item{separator, }{separator for matrix file}

\item{cloutput, }{number of permutations to perform the pValue to evaluate clustering. Suggested minimal number of permutations 10}

\item{perplexity, }{TSNE parameter,  sets the number of nearest neighbors to consider when constructing a probability distribution over the input data. A higher perplexity value will result in a more accurate probability distribution, while a lower value will result in a simpler distribution that is faster to compute.}

\item{n_iter, }{TSNE parameter, This parameter sets the number of iterations to run the optimization process. A higher number of iterations will result in a more accurate solution, but will also take longer to compute. A lower number of iterations will result in a faster computation, but may not find the optimal solution}

\item{n_neighbors, }{UMAP parameter, sets the number of neighbors used in the local neighborhood graph construction. A higher value of n_neighbors will result in a more complex and accurate neighborhood graph, while a lower value will result in a simpler graph that is faster to compute.}

\item{min_dist, }{UMAP parameter, sets the minimum distance between any two points in the low-dimensional embedding. A higher value of min_dist will result in a more spread-out embedding, while a lower value will result in a more compact embedding. Min_dist can be used to control the clustering of points in the embedding, with smaller values leading to tighter clusters and larger values leading to more dispersed clusters.}

\item{tsneplot, }{1 for tsne plot 0 otherwise}

\item{umapplot, }{1 for umap plot 0 otherwise}

\item{matrixName, }{a character string indicating the path of the file, with file name and extension included}
}
\description{
The present function compress data using autoencoder partially connected creating pseudoBulk matrix
}
\examples{
\dontrun{
}
}
\author{
Luca Alessandri, alessandri [dot] luca1991 [at] gmail [dot] com, University of Torino
}
